+++ src/main/kotlin/com/codescene/jetbrains/listeners/ProjectStartupActivity.kt
@@ -5,6 +5,9 @@ import com.codescene.jetbrains.config.global.CodeSceneGlobalSettingsStore
 import com.codescene.jetbrains.util.Constants.CODESCENE
 import com.codescene.jetbrains.util.Log
 import com.codescene.jetbrains.util.showNotification
+import com.intellij.ide.plugins.IdeaPluginDescriptor
+import com.intellij.ide.plugins.PluginInstaller
+import com.intellij.ide.plugins.PluginStateListener
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.Project
@@ -27,6 +30,18 @@ class ProjectStartupActivity : ProjectActivity {
             NotificationType.INFORMATION
         )
 
+        addStateListener()
         VirtualFileManager.getInstance().addAsyncFileListener(FileChangeListener(project), disposable)
     }
+
+    private fun addStateListener() = PluginInstaller.addStateListener(object : PluginStateListener {
+        override fun install(descriptor: IdeaPluginDescriptor) {
+            // No action needed
+        }
+
+        override fun uninstall(descriptor: IdeaPluginDescriptor) {
+            Log.info("Plugin uninstalled: ${descriptor.pluginId} ${descriptor.version}")
+            CodeSceneGlobalSettingsStore.getInstance().updateTermsAndConditionsAcceptance(false)
+        }
+    })
 }