+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Notification prompt on project startup for Terms and conditions.
+
+### Changed
+- Review and Delta services don't trigger until Terms and Conditions are accepted.
 
 ## [0.0.8-beta] - 2024-01-27
 ### Changed